Refactor code for readability and separation of concerns
